pyproject.toml
[tool.poetry]
name = "noteslink"
version = "0.1.0"
description = "Efficient note-taking API for seamless management of thoughts and tasks"
authors = ["bhaveshdev09 <patilbhavesh991209@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "4.2.8"
django-environ = "^0.11.2"
djangorestframework = "^3.14.0"
djangorestframework-simplejwt = "^5.3.1"
drf-spectacular = "^0.27.1"


[tool.poetry.group.dev.dependencies]
django-shortcuts = "^1.6"
ipython = "^8.21.0"
coverage = "^7.4.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
